+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 ### Added
 - resource pullzone_hostname: manage free TLS certificate
+### Changed
+- resource pullzone_hostname: import via name instead of ID
 
 ## [0.3.3] - 2024-07-30
 ### Added